Fix operator overload detection (#169)
Fix a regression introduced by #159.  Nested template specializations
need to have their name extracted to be documented correctly, but the
implementation checked for `node.name` having the same number of `<`
and `>`.  This breaks operator overloading (#168).  Skip asserting the
same number of `<` and `>` when `node.kind == "function"` and `operator`
is in `node.name`.  See comments in `graph.py`.